Update Student Performance Prediction.py
Commented out the lines that display the predictions themselves; trying to clean up the code/output.

Checked feature importance and removed most features leaving only "Gender", "Lunch", and "Prep" -- indicating the students' gender, whether they got a free or standard lunch, and whether or not the prepped for the exams.

This increased accuracy for all models.
Logistic Regression is consistently the most accurate model at 78%.

I ran a for loop to determine the best number of trees to use in the random forest -- the number that would produce the highest success rate -- 2 trees proved to be best.

The Random Forest model ranged between 75%-78% accuracy.

The MLR model was the least accurate. Before reducing the number of features the accuracy was 11%, after updating it went up to about 22%.